FIX: delete WriteBoard,WriteComment
